scripts/requirements.txt
umap-learn>=0.5.0
scikit-learn>=1.3.0
bokeh>=3.0.0
numpy>=1.24.0
pillow>=10.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
